ai-companion/scripts/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-genai>=0.7.0
